requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
load-dotenv==0.1.0
MarkupSafe==3.0.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
